+++ src/index.js
@@ -191,8 +191,15 @@ export default class {
     const listDocumentsRead = []
     const listTotalPending = []
 
-    const calculateSyncProgress = (syncState, listDocumentsRead, listTotalPending, collectionIndex, moduleByName) => {
-      syncState.change$.subscribe(({ change, direction }) => {
+    const calculateSyncProgress = (syncState, listDocumentsRead, listTotalPending, collectionIndex, moduleByName, syncOptions) => {
+      syncState.change$.subscribe(context => {
+        // if it is only push there will be no loading.
+        if (syncOptions.direction.push && !syncOptions.direction.pull) return
+
+        // when it's only one direction the structure comes { ...contentChange }
+        context = context.direction ? context : { direction: 'pull', change: { ...context } }
+
+        const { change, direction } = context
         const isPullDirection = direction === 'pull'
 
         if (isPullDirection && this.sync.progress) {
@@ -235,6 +242,7 @@ export default class {
       const moduleByName = this.modules.find(module => module.name === collectionName)
       const moduleOptions = (moduleByName.sync && moduleByName.sync.options) || {}
       const syncOptions = { ...defaultOptions, ...this.sync.options, ...moduleOptions }
+      const query = (moduleByName.sync && moduleByName.sync.query) || this.sync.query || (() => {})
 
       if (!syncOptions.baseURL) {
         throw new Error('baseURL is required to sync.')
@@ -242,7 +250,8 @@ export default class {
 
       const syncState = await this.collections[collectionName].sync({
         ...syncOptions,
-        remote: `${syncOptions.baseURL}/${collectionName}`
+        remote: `${syncOptions.baseURL}/${collectionName}`,
+        query: query(this.collections[collectionName])
       })
 
       if (moduleByName.sync && moduleByName.sync.handler) {
@@ -250,7 +259,7 @@ export default class {
       }
 
       if (this.sync.progress || (moduleByName.sync && moduleByName.sync.progress)) {
-        calculateSyncProgress(syncState, listDocumentsRead, listTotalPending, collectionIndex, moduleByName)
+        calculateSyncProgress(syncState, listDocumentsRead, listTotalPending, collectionIndex, moduleByName, syncOptions)
       }
     }
   }