+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+# Unreleased
+
+### Added
+- added new method `query` to sync.
+
 # 2.4.1 2021-08-19
 
 ### Fixed
-- added new method `deleteBy` for `setDefaults` work correctly
+- added new method `deleteBy` for `setDefaults` work correctly.
 
 # 2.4.0 2021-08-12